exported categories array
